requirements.txt
pydantic
pyinstaller
jpype1